+++ .gitignore
@@ -16,4 +16,14 @@
 
 
 # End of https://www.gitignore.io/api/go
-vendor/
+
+# Serverless directories
+.serverless
+
+# golang output binary directory
+bin
+
+# golang vendor (dependencies) directory
+vendor
+
+node_modules
