+++ Gopkg.toml
@@ -36,3 +36,7 @@
 [[constraint]]
   branch = "master"
   name = "github.com/serjvanilla/go-overpass"
+
+[[constraint]]
+  name = "github.com/aws/aws-lambda-go"
+  version = "^1.0.1"
