+++ Makefile
@@ -1,0 +1,3 @@
+build:
+	dep ensure
+	env GOOS=linux go build -ldflags="-s -w" -o bin/remote remote/main.go
