+++ package.json
@@ -1,0 +1,11 @@
+{
+  "name": "FIT3036-backend",
+  "version": "1.0.0",
+  "main": "index.js",
+  "repository": "git@github.com:dylanpinn/FIT3036-backend.git",
+  "author": "Dylan Pinn <dylan@dylanpinn.com>",
+  "license": "MIT",
+  "devDependencies": {
+    "serverless": "^1.27.2"
+  }
+}
